feat: Rename namespace and enhance code
