+++ lib/RoadizRozierBundle/composer.json
@@ -21,7 +21,8 @@
         "php": ">=8.3",
         "roadiz/core-bundle": "2.6.*",
         "roadiz/openid": "2.6.*",
-        "symfony/framework-bundle": "7.3.*"
+        "symfony/framework-bundle": "7.3.*",
+        "deeplcom/deepl-php": "^1.12"
     },
     "require-dev": {
         "php-coveralls/php-coveralls": "^2.4",
