+++ lib/RoadizRozierBundle/config/routing/translate-assistant.yml
@@ -1,8 +1,8 @@
 translate:
     path: /translate
     defaults:
-        _controller: RZ\Roadiz\RozierBundle\Translator\TranslateAssistantController::translateAction
+        _controller: RZ\Roadiz\RozierBundle\TranslateAssistant\TranslateAssistantController::translateAction
 rephrase:
     path: /rephrase
     defaults:
-        _controller: RZ\Roadiz\RozierBundle\Translator\TranslateAssistantController::rephraseAction
+        _controller: RZ\Roadiz\RozierBundle\TranslateAssistant\TranslateAssistantController::rephraseAction
