+++ lib/RoadizRozierBundle/src/Command/TranslateAssistantRephraseCommand.php
@@ -4,12 +4,11 @@ declare(strict_types=1);
 
 namespace RZ\Roadiz\RozierBundle\Command;
 
-use RZ\Roadiz\RozierBundle\Translator\TranslateAssistantDto;
-use RZ\Roadiz\RozierBundle\Translator\TranslateAssistantInterface;
+use RZ\Roadiz\RozierBundle\TranslateAssistant\TranslateAssistantInput;
+use RZ\Roadiz\RozierBundle\TranslateAssistant\TranslateAssistantInterface;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
 
@@ -32,7 +31,11 @@ final class TranslateAssistantRephraseCommand extends Command
                 InputArgument::REQUIRED,
                 'Text to rephrase.'
             )
-            ->addOption('lang', 'l', InputOption::VALUE_REQUIRED, 'Set locale used to rephrase.')
+            ->addArgument(
+                'lang',
+                InputArgument::REQUIRED,
+                'Set locale used to rephrase.'
+            )
         ;
     }
 
@@ -41,7 +44,7 @@ final class TranslateAssistantRephraseCommand extends Command
     {
         $io = new SymfonyStyle($input, $output);
         $text = $input->getArgument('text');
-        $lang = $input->getOption('lang');
+        $lang = $input->getArgument('lang');
 
         if (!\is_string($text) || empty($text)) {
             throw new \InvalidArgumentException('Text argument is required.');
@@ -51,7 +54,7 @@ final class TranslateAssistantRephraseCommand extends Command
             throw new \InvalidArgumentException('Lang option is required.');
         }
 
-        $dto = new TranslateAssistantDto(
+        $dto = new TranslateAssistantInput(
             $text,
             $lang,
             $lang
