+++ lib/RoadizRozierBundle/src/TranslateAssistant/TranslateAssistantInput.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace RZ\Roadiz\RozierBundle\Translator;
+namespace RZ\Roadiz\RozierBundle\TranslateAssistant;
 
-final readonly class TranslateAssistantDto
+final readonly class TranslateAssistantInput
 {
     public function __construct(
         public string $text,
