--- lib/RoadizRozierBundle/src/Translator/TranslateAssistantInterface.php
@@ -1,12 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace RZ\Roadiz\RozierBundle\Translator;
-
-interface TranslateAssistantInterface
-{
-    public function translate(TranslateAssistantDto $translatorDto): TranslateAssistantOutput;
-
-    public function rephrase(TranslateAssistantDto $translatorDto): TranslateAssistantOutput;
-}
